moves reference point transform method to separate class
