+++ Connectors/Revit/Speckle.Connectors.RevitShared/Bindings/RevitSendBinding.cs
@@ -1,6 +1,5 @@
 using System.Collections.Concurrent;
 using Autodesk.Revit.DB;
-using Autodesk.Revit.UI;
 using Autofac;
 using Microsoft.Extensions.Logging;
 using Speckle.Autofac.DependencyInjection;
@@ -88,6 +87,7 @@ internal sealed class RevitSendBinding : RevitBaseBinding, ISendBinding
   // cache invalidation process run with ModelCardId since the settings are model specific
   private readonly Dictionary<string, DetailLevelType> _detailLevelCache = new();
   private readonly Dictionary<string, ReferencePointType> _referencePointCache = new();
+  private readonly ReferencePointManager _referencePointManager = new();
 
   private ToSpeckleSettings GetToSpeckleSettings(SenderModelCard modelCard)
   {
@@ -139,62 +139,9 @@ internal sealed class RevitSendBinding : RevitBaseBinding, ISendBinding
           _sendConversionCache.EvictObjects(objectIds);
         }
       }
+
       _referencePointCache[modelCard.ModelCardId.NotNull()] = referencePoint;
-
-      // Get the doc transform from reference point setting
-      Transform? referencePointTransform = null;
-
-      // first get the main doc base points and reference setting transform
-      if (RevitContext.UIApplication is UIApplication uiApplication)
-      {
-        using FilteredElementCollector filteredElementCollector = new(uiApplication.ActiveUIDocument.Document);
-        var points = filteredElementCollector.OfClass(typeof(BasePoint)).Cast<BasePoint>().ToList();
-        BasePoint? projectPoint = points.FirstOrDefault(o => !o.IsShared);
-        BasePoint? surveyPoint = points.FirstOrDefault(o => o.IsShared);
-
-        switch (referencePoint)
-        {
-          // note that the project base (ui) rotation is registered on the survey pt, not on the base point
-          case ReferencePointType.ProjectBase:
-            if (projectPoint is not null)
-            {
-              referencePointTransform = Transform.CreateTranslation(projectPoint.Position);
-            }
-            else
-            {
-              throw new InvalidOperationException("Couldn't retrieve Project Point from document");
-            }
-            break;
-
-          // note that the project base (ui) rotation is registered on the survey pt, not on the base point
-          case ReferencePointType.Survey:
-            if (surveyPoint is not null && projectPoint is not null)
-            {
-              // POC: should a null angle resolve to 0?
-              // retrieve the survey point rotation from the project point
-              var angle = projectPoint.get_Parameter(BuiltInParameter.BASEPOINT_ANGLETON_PARAM)?.AsDouble() ?? 0;
-
-              // POC: following disposed incorrectly or early or maybe a false negative?
-              using Transform translation = Transform.CreateTranslation(surveyPoint.Position);
-              referencePointTransform = translation.Multiply(Transform.CreateRotation(XYZ.BasisZ, angle));
-            }
-            else
-            {
-              throw new InvalidOperationException("Couldn't retrieve Survey and Project Point from document");
-            }
-            break;
-
-          case ReferencePointType.InternalOrigin:
-            break;
-
-          default:
-            break;
-        }
-
-        return referencePointTransform;
-      }
-
-      throw new InvalidOperationException("Revit Context UI Application was null");
+      return _referencePointManager.GetTransform(RevitContext, referencePoint);
     }
 
     throw new ArgumentException($"Invalid reference point value: {referencePointString}");
