+++ Connectors/Revit/Speckle.Connectors.RevitShared/Operations/Send/Settings/ReferencePointSetting.cs
@@ -1,4 +1,7 @@
+using Autodesk.Revit.DB;
+using Autodesk.Revit.UI;
 using Speckle.Connectors.DUI.Settings;
+using Speckle.Converters.RevitShared.Helpers;
 using Speckle.Converters.RevitShared.Settings;
 
 namespace Speckle.Connectors.Revit.Operations.Send.Settings;
@@ -15,4 +18,66 @@ public class ReferencePointSetting(ReferencePointType value) : ICardSetting
     .Enum.GetValues(typeof(ReferencePointType))
     .Cast<ReferencePointType>()
     .ToDictionary(v => v.ToString(), v => v);
+}
+
+public class ReferencePointManager()
+{
+  public Transform? GetTransform(RevitContext context, ReferencePointType referencePointType)
+  {
+    Transform? referencePointTransform = null;
+
+    if (context.UIApplication is UIApplication uiApplication)
+    {
+      // first get the main doc base points and reference setting transform
+      using FilteredElementCollector filteredElementCollector = new(uiApplication.ActiveUIDocument.Document);
+      var points = filteredElementCollector.OfClass(typeof(BasePoint)).Cast<BasePoint>().ToList();
+      BasePoint? projectPoint = points.FirstOrDefault(o => !o.IsShared);
+      BasePoint? surveyPoint = points.FirstOrDefault(o => o.IsShared);
+
+      switch (referencePointType)
+      {
+        // note that the project base (ui) rotation is registered on the survey pt, not on the base point
+        case ReferencePointType.ProjectBase:
+          if (projectPoint is not null)
+          {
+            referencePointTransform = Transform.CreateTranslation(projectPoint.Position);
+          }
+          else
+          {
+            throw new InvalidOperationException("Couldn't retrieve Project Point from document");
+          }
+          break;
+
+        // note that the project base (ui) rotation is registered on the survey pt, not on the base point
+        case ReferencePointType.Survey:
+          if (surveyPoint is not null && projectPoint is not null)
+          {
+            // POC: should a null angle resolve to 0?
+            // retrieve the survey point rotation from the project point
+            var angle = projectPoint.get_Parameter(BuiltInParameter.BASEPOINT_ANGLETON_PARAM)?.AsDouble() ?? 0;
+
+            // POC: following disposed incorrectly or early or maybe a false negative?
+            using Transform translation = Transform.CreateTranslation(surveyPoint.Position);
+            referencePointTransform = translation.Multiply(Transform.CreateRotation(XYZ.BasisZ, angle));
+          }
+          else
+          {
+            throw new InvalidOperationException("Couldn't retrieve Survey and Project Point from document");
+          }
+          break;
+
+        case ReferencePointType.InternalOrigin:
+          break;
+
+        default:
+          break;
+      }
+
+      return referencePointTransform;
+    }
+
+    throw new InvalidOperationException(
+      "Revit Context UI Application was null when retrieving reference point transform."
+    );
+  }
 }
